requirements.txt
boto3
requests
pandas
tqdm
flamingo-pytorch
decord
pillow
einops
easydict
timm
opencv-python
flow_vis
h5py
av
tensorboard
terminaltables
fvcore
lz4
regex
colorlog
tabulate
scikit-learn